Move online/announcement into block dir
